pyproject.toml
[project]
name = "dlab"
version = "0.1.0"
description = "A local-first deep learning experimentation OS for rapid research loops."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.0",
    "hydra-core>=1.3.2",
    "jupyterlab>=4.2.0",
    "jupytext>=1.16.0",
    "litlogger>=2026.5.12",
    "matplotlib>=3.9.0",
    "numpy>=2.0.0",
    "omegaconf>=2.3.0",
    "pandas>=2.2.0",
    "pillow>=10.4.0",
    "pytorch-lightning>=2.4.0",
    "scikit-learn>=1.5.0",
    "timm>=1.0.0",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "transformers>=4.44.0",
    "wandb>=0.17.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
